+++ android/src/main/java/com/lib/flutter_blue_plus/FlutterBluePlusPlugin.java
@@ -103,7 +103,7 @@ public class FlutterBluePlusPlugin implements
     private final Map<String, BluetoothGatt> mCurrentlyConnectingDevices = new ConcurrentHashMap<>();
     private final Map<String, BluetoothDevice> mBondingDevices = new ConcurrentHashMap<>();
     private final Map<String, Integer> mMtu = new ConcurrentHashMap<>();
-    private final Map<String, Boolean> mAutoConnected = new ConcurrentHashMap<>();
+    private final Map<String, BluetoothGatt> mAutoConnected = new ConcurrentHashMap<>();
     private final Map<String, String> mWriteChr = new ConcurrentHashMap<>();
     private final Map<String, String> mWriteDesc = new ConcurrentHashMap<>();
     private final Map<String, String> mAdvSeen = new ConcurrentHashMap<>();
@@ -265,7 +265,7 @@ public class FlutterBluePlusPlugin implements
     //  ██████  ██   ██  ███████  ███████
 
     @Override
-    @SuppressWarnings({"deprecation", "unchecked"}) // needed for compatability, type safety uses bluetooth_msgs.dart
+    @SuppressWarnings({"deprecation", "unchecked"}) // needed for compatibility, type safety uses bluetooth_msgs.dart
     public void onMethodCall(@NonNull MethodCall call,
                                  @NonNull Result result)
     {
@@ -703,7 +703,7 @@ public class FlutterBluePlusPlugin implements
 
                         // remember autoconnect 
                         if (autoConnect) {
-                            mAutoConnected.put(remoteId, autoConnect);
+                            mAutoConnected.put(remoteId, gatt);
                         } else {
                             mAutoConnected.remove(remoteId);
                         }
@@ -727,6 +727,17 @@ public class FlutterBluePlusPlugin implements
                     }
                     if (gatt == null) {
                         gatt = mConnectedDevices.get(remoteId);;
+                    }
+                    if (gatt == null) {
+                        gatt = mAutoConnected.get(remoteId);
+                        if (gatt != null) {
+                            log(LogLevel.DEBUG, "already disconnected. disabling autoconnect");
+                            mAutoConnected.remove(remoteId);
+                            gatt.disconnect();
+                            gatt.close();
+                            result.success(false);  // no work to do
+                            return;
+                        }
                     }
                     if (gatt == null) {
                         log(LogLevel.DEBUG, "already disconnected");
@@ -1714,6 +1725,60 @@ public class FlutterBluePlusPlugin implements
         mAutoConnected.clear();
     }
 
+    int getAppearanceFromScanRecord(ScanRecord adv) {
+
+        if (Build.VERSION.SDK_INT >= 33) { // Android 13 (August 2022)
+            Map<Integer, byte[]> map = adv.getAdvertisingDataMap();
+            if (map.containsKey(ScanRecord.DATA_TYPE_APPEARANCE)) {
+                byte[] bytes = map.get(ScanRecord.DATA_TYPE_APPEARANCE);
+                if (bytes.length == 2) {
+                    int loByte = bytes[0] & 0xFF;
+                    int hiByte = bytes[1] & 0xFF;
+                    return hiByte * 256 + loByte;
+                }
+            }
+            return 0;
+        }
+
+        // For API Level 21+
+        byte[] bytes = adv.getBytes();
+
+        int n = 0;
+
+        while (n < bytes.length) {
+
+            int fieldLen = bytes[n];
+
+            // no more or malformed data
+            if (fieldLen <= 0) {
+                break;
+            }
+
+            // end of packet
+            if (fieldLen + n > bytes.length - 1) {
+                break;
+            }
+
+            int dataType = bytes[n + 1];
+
+            // no more data
+            if (dataType == 0) {
+                break;
+            }
+
+            // appearance type byte
+            if (dataType == 0x19 && fieldLen == 3) {
+                int loByte = bytes[n + 2] & 0xFF;
+                int hiByte = bytes[n + 3] & 0xFF;
+                return hiByte * 256 + loByte;
+            }
+
+            n += fieldLen + 1;
+        }
+
+        return 0;
+    }
+
     /////////////////////////////////////////////////////////////////////////////////////
     //  █████   ██████    █████   ██████   ████████  ███████  ██████
     // ██   ██  ██   ██  ██   ██  ██   ██     ██     ██       ██   ██
@@ -2295,6 +2360,7 @@ public class FlutterBluePlusPlugin implements
 
         String                  advName      = adv != null ?  adv.getDeviceName()                : null;
         int                     txPower      = adv != null ?  adv.getTxPowerLevel()              : min;
+        int                     appearance   = adv != null ?  getAppearanceFromScanRecord(adv)   : 0;
         SparseArray<byte[]>     manufData    = adv != null ?  adv.getManufacturerSpecificData()  : null;
         List<ParcelUuid>        serviceUuids = adv != null ?  adv.getServiceUuids()              : null;
         Map<ParcelUuid, byte[]> serviceData  = adv != null ?  adv.getServiceData()               : null;
@@ -2335,6 +2401,7 @@ public class FlutterBluePlusPlugin implements
         if (connectable)                 {map.put("connectable", 1);}
         if (advName != null)             {map.put("adv_name", advName);}
         if (txPower != min)              {map.put("tx_power_level", txPower);}
+        if (appearance != 0)             {map.put("appearance", appearance);}
         if (manufData != null)           {map.put("manufacturer_data", manufDataB);}
         if (serviceData != null)         {map.put("service_data", serviceDataB);}
         if (serviceUuids != null)        {map.put("service_uuids", serviceUuidsB);}