+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.31.17
+* **[Fix]** android: autoconnect could enter disconnect / connection loop
+* **[Fix]** android: calling `disconnect` should always disable autoconnect, even if already disconnected
+
+## 1.31.16
+* **[Fix]** forgot to push to mtu stream after disconnection
+
+## 1.31.15
+* **[Feature]** support advertising `appearance` on Android
+
+## 1.31.14
+* **[Fix]** autoconnect could not disconnect if there are multiple devices
+
 ## 1.31.13
 * **[Improve]**  android: use name `LINK_SUPERVISION_TIMEOUT` instead of `CONNECTION_TIMEOUT`